requirements.txt
python-telegram-bot[job-queue]==20.1
pytz==2022.7.1
requests==2.28.2
schedule==1.1.0
